backend/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==3.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
chardet==3.0.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmd2==1.5.0
colorama==0.4.6
commonmark==0.9.1
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
fsspec==2024.12.0
ftfy==6.3.1
h11==0.14.0
huggingface-hub==0.27.0
idna==2.10
Jinja2==3.0.0
joblib==1.4.2
kombu==5.4.2
Markdown==3.7
markdownify==0.13.1
marker-pdf
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx
numpy
opencv-python==4.10.0.84
packaging==24.2
pdftext
pillow==10.4.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.11.0
redis==5.2.1
regex==2024.11.6
requests==2.25.1
rich==10.2.0
ruamel.yaml==0.17.17
safetensors==0.4.5
scikit-learn==1.6.0
scipy
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
surya-ocr
sympy
tabled-pdf
tabulate==0.9.0
texify
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl
